// Courses-platform/src/app/core/auth/models/decoded-token.interface.ts
export interface IDecodedToken {
    Id:                                                                     string;
    "http://schemas.xmlsoap.org/ws/2005/05/identity/claims/name":           string;
    "http://schemas.xmlsoap.org/ws/2005/05/identity/claims/emailaddress":   string;
    "http://schemas.xmlsoap.org/ws/2005/05/identity/claims/nameidentifier": string;
    "http://schemas.microsoft.com/ws/2008/06/identity/claims/expiration":   string;
    "http://schemas.microsoft.com/ws/2008/06/identity/claims/role":         string;
    nbf:                                                                    number;
    exp:                                                                    number;
    iss:                                                                    string;
    aud:                                                                    string;
}